+++ test/TestOpCreateTensor.cpp
@@ -9,7 +9,7 @@ TEST_CASE("test_opcreatetensor_create_single_tensor") {
 
     kp::Manager mgr;
 
-    std::vector<uint32_t> testVecA{ 9, 8, 7 };
+    std::vector<float> testVecA{ 9, 8, 7 };
 
     std::shared_ptr<kp::Tensor> tensorA{new kp::Tensor(testVecA)};
 
@@ -27,8 +27,8 @@ TEST_CASE("test_opcreatetensor_create_multiple_tensors_single_op") {
 
     kp::Manager mgr;
 
-    std::vector<uint32_t> testVecA{ 9, 8, 7 };
-    std::vector<uint32_t> testVecB{ 6, 5, 4 };
+    std::vector<float> testVecA{ 9, 8, 7 };
+    std::vector<float> testVecB{ 6, 5, 4 };
 
     std::shared_ptr<kp::Tensor> tensorA{new kp::Tensor(testVecA)};
     std::shared_ptr<kp::Tensor> tensorB{new kp::Tensor(testVecB)};
@@ -46,8 +46,8 @@ TEST_CASE("test_opcreatetensor_create_multiple_tensors_multiple_op") {
 
     kp::Manager mgr;
 
-    std::vector<uint32_t> testVecA{ 9, 8, 7 };
-    std::vector<uint32_t> testVecB{ 6, 5, 4 };
+    std::vector<float> testVecA{ 9, 8, 7 };
+    std::vector<float> testVecB{ 6, 5, 4 };
 
     std::shared_ptr<kp::Tensor> tensorA{new kp::Tensor(testVecA)};
     std::shared_ptr<kp::Tensor> tensorB{new kp::Tensor(testVecB)};
@@ -64,8 +64,8 @@ TEST_CASE("test_opcreatetensor_create_multiple_tensors_multiple_op") {
 
 TEST_CASE("test_opcreatetensor_manage_tensor_memory_when_destroyed") {
 
-    std::vector<uint32_t> testVecA{ 9, 8, 7 };
-    std::vector<uint32_t> testVecB{ 6, 5, 4 };
+    std::vector<float> testVecA{ 9, 8, 7 };
+    std::vector<float> testVecB{ 6, 5, 4 };
 
     std::shared_ptr<kp::Tensor> tensorA{new kp::Tensor(testVecA)};
     std::shared_ptr<kp::Tensor> tensorB{new kp::Tensor(testVecB)};
@@ -88,8 +88,8 @@ TEST_CASE("test_opcreatetensor_manage_tensor_memory_when_destroyed") {
 
 TEST_CASE("test_opcreatetensor_no_error_if_tensor_freed_before") {
 
-    std::vector<uint32_t> testVecA{ 9, 8, 7 };
-    std::vector<uint32_t> testVecB{ 6, 5, 4 };
+    std::vector<float> testVecA{ 9, 8, 7 };
+    std::vector<float> testVecB{ 6, 5, 4 };
 
     std::shared_ptr<kp::Tensor> tensorA{new kp::Tensor(testVecA)};
     std::shared_ptr<kp::Tensor> tensorB{new kp::Tensor(testVecB)};
